Extract HelpMenu class
